requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
multidict==6.0.4
mysqlclient==2.2.1
openai==1.6.1
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.1.0
PyJWT==2.8.0
pyppeteer==2.0.0
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
websockets==10.4
yarl==1.9.4
zipp==3.17.0
